+++ src/render.h
@@ -8,8 +8,9 @@
 
 int16_t windowSize(int8_t axis);
 void cleanFrame(void);
-void lineLenCheck(int8_t chars);
+int8_t lineLenCheck(int8_t chars);
 void window(char key, char filename[32], int8_t chars, int8_t lines);
+void windowEmpty(char filename[32], int8_t chars, int8_t lines);
 
 #endif